Introducing multiplexing oracle instance monitoring :
Move monitoring logic from main.go to exporter/exporter.go.
Add new option --config-dir to specify where we store config file.
Examples of config file in config-example.
