+++ README.md
@@ -48,7 +48,7 @@ Pre-compiled versions for Linux 64 bit and Mac OSX 64 bit can be found under [re
 
 In order to run, you'll need the [Oracle Instant Client Basic](http://www.oracle.com/technetwork/database/features/instant-client/index-097480.html) for your operating system. Only the basic version is required for execution.
 
-# Running the Binary
+# Running the Binary with DATA_SOURCE_NAME variable
 
 Ensure that the environment variable DATA_SOURCE_NAME is set correctly before starting. For Example
 
@@ -57,18 +57,64 @@ export DATA_SOURCE_NAME=system/oracle@myhost
 /path/to/binary -l log.level error -l web.listen-address 9161
 ```
 
+# Running the Binary with config file
+
+Another way to launch the binary is to use a directory where you store your configuration in **poml** file. If you provide more than one file, this exporter load them all. It's a convenient way to multiplex your Oracle monitoring when you've got more than one Oracle instance on a server.
+
+There's two examples in ``config-example``. In this file, you can provide each values on separate fields like this:
+
+```
+user="system"
+pass="temppasswd"
+host="localhost"
+port="1521"
+service="XE"
+```
+
+**host** and **port** are not required and use default value (localhost and 1521). You can change this default value using --listener-address and --listener-port options.
+
+You can also provide only string field with all informations:
+
+```
+string="system/temppasswd@localhost:1521/XE"
+```
+
+By default, this exporter add a name label using the filename (without the extension). If you want to change this label, you can use **name** field with the value you wan't:
+
+```
+name="DEMO"
+```
+
+Here is an example using one config file for **oracledb_sessions_active**:
+
+```
+# HELP oracledb_sessions_active Gauge metric with count of sessions marked ACTIVE
+# TYPE oracledb_sessions_active gauge
+oracledb_sessions_active{name="XE"} 41
+```
+
+Same metrics with two config file (XE.poml and DEMO.poml):
+```
+# HELP oracledb_sessions_active Gauge metric with count of sessions marked ACTIVE
+# TYPE oracledb_sessions_active gauge
+oracledb_sessions_active{name="DEMO"} 41
+oracledb_sessions_active{name="XE"} 31
+```
+
 ## Usage
 
 ```bash
 Usage of oracledb_exporter:
-  -log.format value
-       	If set use a syslog logger or JSON logging. Example: logger:syslog?appname=bob&local=7 or logger:stdout?json=true. Defaults to stderr.
-  -log.level value
-       	Only log messages with the given severity or above. Valid levels: [debug, info, warn, error, fatal].
+  -config-dir string
+        Directory where we store config. Each file result in map
+  -listener-address string
+        Default Oracle listener address (default "localhost")
+  -listener-port string
+        Default Oracle listener port (default "1521")
   -web.listen-address string
-       	Address to listen on for web interface and telemetry. (default ":9161")
+        Address to listen on for web interface and telemetry. (default ":9161")
   -web.telemetry-path string
-       	Path under which to expose metrics. (default "/metrics")
+        Path under which to expose metrics. (default "/metrics")
 ```
 
 # Integration with Grafana
